backend/pyproject.toml
[project]
name = "energy-efficiency-tracker-api"
version = "0.1.0"
description = "Energy Efficiency Tracker API"
authors = [
    {name = "Maykol Laguna", email = "lagunamichael2@outlook.com"},
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.0.0",
    "redis>=6.4.0",
    "structlog>=25.4.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
lint = [
    "ruff>=0.13.3",
]
